pyproject.toml
[tool.poetry]
name = "irises-classification"
version = "0.1.0"
description = ""
authors = ["ksushapolo"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
argparse = "^1.4.0"
pandas = "^2.0.0"
dvc = "^3.30.3"
dvc-gdrive = "^2.20.0"
hydra = "^2.5"
mlflow = "^2.8.1"
gitpython = "^3.1.40"

[tool.poetry.group.dev.dependencies]
scikit-learn = "^1.3.1"
pre-commit = "^3.4.0"

[tool.black]
line-length = 90
target-version = ["py38"]
skip-string-normalization = true

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 90
lines_after_imports = 2

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
